Run QA tools

+++ app/Console/Commands/SelectPhotosForTraining.php
@@ -173,7 +173,7 @@ class SelectPhotosForTraining extends Command
         $units = ['B', 'KB', 'MB', 'GB', 'TB', 'PB'];
 
         $bytes = max($bytes, 0);
-        $pow = floor(($bytes ? log($bytes) : 0) / log(1024));
+        $pow = (int) floor(($bytes ? log($bytes) : 0) / log(1024));
         $pow = min($pow, count($units) - 1);
 
         $bytes /= 1024 ** $pow;
